pyproject.toml
[project]
name = "imagebot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiogram>=3.19.0",
    "django>=5.2",
    "openpyxl>=3.1.5",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.4",
    "xlwt>=1.3.0",
]
